Incorporating some of dimitarvdimitrov's feedback
